Support the --proxy option
The reqwest crate does all the heavy lifting (including using the
http_proxy and https_proxy environment variables when the --proxy option
isn't present).
As noted, it's possible to use different proxies for different
protocols.

The "socks" configuration for the reqwest crate was enabled to allow
users to use a socks proxy.