Replace http2_adaptive_window by fixed window to prevent crashes
The HTTP/2 adaptive window feature sends a lot of pings. Cloudflare
doesn't like that and eventually breaks off the connection.

We can just set a large initial window instead. A window of 4 MiB
appears to give the same speed (tested on Linux with a local HTTP/2
server and an artificial delay).

This way we also reach maximum speed earlier, the adaptive window
takes a while to grow.

I'm unsure how this affects memory usage, particularly on Windows,
which doesn't overcommit memory.