Switch to full buffering and manual flushing
All output is now fully buffered. That means it might not show up
until `.flush()` is called, so `.flush()` has to be called manually
pretty frequently. The general rule is that it should be called when
the next output might not be ready yet. For example, `.flush()` is
called at the end of `print_request_body()`, so the request is visible
before waiting for the response.

This gets trickiest for streaming responses. They have to flush after
receiving the next piece of response from the server, but not more
often than that (or it'd hurt performance). `BinaryGuard` and
`copy_largebuf` have been changed to make that possible.

But these are also where the greatest gains are made: A piped
`--stream --pretty format` JSON response is now ten times as
fast as before. That's admittedly a bit niche, but I think `--stream`
might now always be faster than not streaming. (I haven't benchmarked
it exhaustively yet.)

The `Buffer` type is overhauled, in part for unrelated
optimization. It turns out that it can matter a lot which specific
types from `termcolor` are used.