Use a proper error type for binary data detection
We used to use `ErrorKind::InvalidData` to communicate binary data
that should not be shown in the terminal but that one can actually
happen in other cases as well. brotli decoding uses that ErrorKind,
and we now use it for all decompressors.

So an invalid brotli response would under certain circumstances render
as "NOTE: binary data not shown in terminal".

We can use our own error type to track this properly.